Revert "fix vs2017"
This reverts commit 80177bba5f642195fdf35f86a7701dd5a03f1d96.

+++ cocos/bindings/manual/jsb_scene_manual.cpp
@@ -630,10 +630,10 @@ static bool js_scene_Node_rotateForJS(void *s) // NOLINT(readability-identifier-
         cobj->rotateForJS(tempFloatArray[1], tempFloatArray[2], tempFloatArray[3], tempFloatArray[4]);
     } else {
         auto size = static_cast<int32_t>(tempFloatArray[5]);
-        cobj->rotateForJS(tempFloatArray[1], tempFloatArray[2], tempFloatArray[3], tempFloatArray[4], size == 0 ? cc::NodeSpace::LOCAL : static_cast<cc::NodeSpace>(static_cast<int>(std::roundf(tempFloatArray[5]))));
+        cobj->rotateForJS(tempFloatArray[1], tempFloatArray[2], tempFloatArray[3], tempFloatArray[4], size == 0 ? cc::NodeSpace::LOCAL : static_cast<cc::NodeSpace>(std::roundf(tempFloatArray[5])));
     }
 
-    const auto &lrot  = cobj->getRotation();
+    const auto& lrot = cobj->getRotation();
     tempFloatArray[0] = lrot.x;
     tempFloatArray[1] = lrot.y;
     tempFloatArray[2] = lrot.z;
